remove string reading functions which depend on charset
